% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_granularity_levels_up.R
\name{add_granularity_levels_up}
\alias{add_granularity_levels_up}
\title{Function add data to higher granularity levels}
\usage{
add_granularity_levels_up(df, list_ontologies, list_dimensions = NULL)
}
\arguments{
\item{df}{a data.frame where the values are included}

\item{list_ontologies}{a character vector with ontologies to be summed up to.}
}
\description{
This function generates values for higher granularity levels.
At the moment it takes A.ADM3 and adds them up to A.ADM2 and A.ADM1 and CH.
}
