% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{alert_success_grey}
\alias{alert_success_grey}
\title{alert with text color grey}
\usage{
alert_success_grey(input)
}
\arguments{
\item{input}{input text string}
}
\description{
alert with text color grey
}
\seealso{
Other Style: 
\code{\link{alert_info_grey}()}
}
\concept{Style}
