% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_px.R
\name{browse_px}
\alias{browse_px}
\title{Function to quickly browse a px-file}
\usage{
browse_px(px_number)
}
\arguments{
\item{px-number}{a character with the px code}
}
\description{
Function to quickly browse a px-file
}
