% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_changes_in_input_file.R
\name{check_changes_in_input_file}
\alias{check_changes_in_input_file}
\title{Function to check changes}
\usage{
check_changes_in_input_file(file_name)
}
\arguments{
\item{file_name}{a character with the file name.}
}
\description{
This function checks whether a (downloaded) file changed through md5-checksum.
}
