% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_helpers.R
\name{check_if_all_languages}
\alias{check_if_all_languages}
\title{check if all languages are present}
\usage{
check_if_all_languages(languages, additional_languages, ignore)
}
\arguments{
\item{languages}{existing languages in the file}

\item{additional_languages}{languages added by the parameter additional_languages}

\item{ignore}{wheter to ignore additional languages or not}
}
\description{
check if all languages are present
}
