% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_and_write_per_unit.R
\name{convert_and_write_per_unit}
\alias{convert_and_write_per_unit}
\title{Function to convert into per capita and per other}
\usage{
convert_and_write_per_unit(
  df,
  file_name,
  how_many = 1000,
  type = "pop",
  simplify_date = F
)
}
\arguments{
\item{df}{a data.frame containing the data to be converted}

\item{file_name}{a character specifiying the name of the file (without path).}

\item{how_many}{by how many of the object. For example by 1000 inhabitants then how_many=1000. For per capita you have to set how_many=1.}

\item{type}{like "pop" or "area" (other types give error out)}

\item{simplify_date}{a bool that simplifies the year in case that it is not an exact match, e.g. 31.12.2020 simply to 2020.}
}
\description{
This function converts e.g. into per capita and then writes the result
}
