% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_cantonal_name_to_current_id.R
\name{convert_cantonal_name_to_current_id}
\alias{convert_cantonal_name_to_current_id}
\title{Function to convert a cantonal name to current id}
\usage{
convert_cantonal_name_to_current_id(kanton, type = "hard")
}
\arguments{
\item{kanton}{a character vector containing the names of the cantons}

\item{type}{either "hard" or "easy" with standard "hard". If hard, then it calculates the Levenshtein distanec.
If easy it assumes that the cantons are already correctly ordered}
}
\description{
This function converts cantonal names to current_ids. Only for cantons!
}
