% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_current_to_hist_id.R
\name{convert_current_to_hist_id}
\alias{convert_current_to_hist_id}
\title{Function to scrape the asset number of a specific bfs number}
\usage{
convert_current_to_hist_id(df, reference_point)
}
\arguments{
\item{df}{a data.frame containing on the one hand an ontology vector called spatialunit_ontology,
and on the other hand a vector called spatialunit_current_id. The latter could be a vector of e.g. bfs-nr for commune, district or canton or any other defined current-id.}

\item{reference_point}{a character date-value in the format \%d.\%m.\%Y to match the spatialunits version.}
}
\description{
This function converts current_ids in combination with ontologies to hist_ids according to spatial-unit table
}
