% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_classes.R
\name{create_classes}
\alias{create_classes}
\title{population_density': '1_01_002',
foreigner_proportion': '1_01_003',}
\usage{
create_classes()
}
\description{
population_density': '1_01_002',
foreigner_proportion': '1_01_003',
}
