% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dimensions.R
\name{create_full_dimension_table}
\alias{create_full_dimension_table}
\title{Function that returns a full dimension table}
\usage{
create_full_dimension_table()
}
\description{
Function that returns a full dimension table
}
