% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_helpers.R
\name{create_new_codes}
\alias{create_new_codes}
\title{create new codes for dimenstion values}
\usage{
create_new_codes(name, value_list)
}
\arguments{
\item{name}{name of the dimension}

\item{value_list}{list with the dimension values}
}
\description{
create new codes for dimenstion values
}
