% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dimensions.R
\name{create_new_dimension_file}
\alias{create_new_dimension_file}
\title{Function that checks and creats a dimension file}
\usage{
create_new_dimension_file(data, overwrite = F, allow_na_cols = F)
}
\arguments{
\item{data}{dimension data}

\item{overwrite}{Whether an existing dimension file should be overwritten or not, default = F}

\item{allow_na_cols}{Whether na-cols are allowed or not (certain columns must have values anyways), default = F}
}
\description{
Function that checks and creats a dimension file
}
