% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dimensions.R
\name{dimension_table_col_names}
\alias{dimension_table_col_names}
\title{function containing all the columns in the right order}
\usage{
dimension_table_col_names()
}
\description{
function containing all the columns in the right order
}
