% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dimensions.R
\name{does_dimension_exist}
\alias{does_dimension_exist}
\title{Function to check if a dimension already exists}
\usage{
does_dimension_exist(uniqe_dimension_name)
}
\arguments{
\item{unique_dimension_name}{Dimension name to test for existance}
}
\description{
Function to check if a dimension already exists
}
