% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bfs_asset_nr.R
\name{download_bfs_file}
\alias{download_bfs_file}
\title{download file without proxy}
\usage{
download_bfs_file(download_url, destfile)
}
\arguments{
\item{download_url}{url to download the file from}

\item{destfile}{destination file path}
}
\description{
download file without proxy
}
