% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dataset.R
\name{download_dataset}
\alias{download_dataset}
\title{Function to download a dataset with a certain dataset_id}
\usage{
download_dataset(dataset_id, flag_force_update)
}
\arguments{
\item{dataset_id}{Id of the dataset to be downloaded}

\item{flag_force_update}{Whether to force the download or not}
}
\description{
Function to download a dataset with a certain dataset_id
}
