% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_helpers.R
\name{extract_dim_value}
\alias{extract_dim_value}
\title{extract the dimension names or the dimenstion values}
\usage{
extract_dim_value(input_list, typ)
}
\arguments{
\item{input_list}{px file list}

\item{typ}{what to extract. either "dim" or "value"}
}
\description{
extract the dimension names or the dimenstion values
}
