% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_helpers.R
\name{extract_languages}
\alias{extract_languages}
\title{extract all existing languages}
\usage{
extract_languages(df_px)
}
\arguments{
\item{df_px}{px-table}
}
\description{
extract all existing languages
}
