% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_meta_and_generate_dimensions.R
\name{extract_meta_and_generate_dimensions}
\alias{extract_meta_and_generate_dimensions}
\title{Function to extract metadata and generate dimensions}
\usage{
extract_meta_and_generate_dimensions(
  input_df,
  unique_names,
  ignore_language,
  additional_languages,
  fuzzy_column_name,
  overwrite
)
}
\arguments{
\item{input_df}{a px-object before conversion into a proper df}

\item{unique_names}{names of the dimenstions to be exportet (in english)}

\item{ignore_language}{if missing languages should be ignored}

\item{additional_languages}{additional language informations as list}

\item{fuzzy_column_name}{name of the column with mixed dimensions}

\item{overwrite}{if an existing dimenstion file should be overwritten or not}
}
\description{
This function extracts metadata from a px-file and populates the dimensions-file
}
