% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_helpers.R
\name{extract_needed_names}
\alias{extract_needed_names}
\title{extract the needed dimension names}
\usage{
extract_needed_names(x, ignore_dim)
}
\arguments{
\item{x}{list with the dimensions}

\item{ignore_dim}{dimenstions to be ignored}
}
\description{
extract the needed dimension names
}
