% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_helpers.R
\name{extract_needed_values}
\alias{extract_needed_values}
\title{extract the list entries with a specific name}
\usage{
extract_needed_values(x_list, needed_names)
}
\arguments{
\item{x_list}{list with the dimension information}

\item{needed_names}{names of list entries that need to be extracted}
}
\description{
extract the list entries with a specific name
}
