% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_spatial_reference.R
\name{extract_spatial_reference}
\alias{extract_spatial_reference}
\title{Function to extract spatial reference from metadata}
\usage{
extract_spatial_reference(char_input)
}
\arguments{
\item{char_input}{a character vector containing the metadata string}
}
\description{
This function extracts the value of the spatial reference that is mentioned in the px-metadata.
It is based on many assumptions and can thus be very buggy! Careful!
}
