% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_helpers.R
\name{extract_value_names}
\alias{extract_value_names}
\title{Extract all existing dimension names}
\usage{
extract_value_names(df_px)
}
\arguments{
\item{df_px}{px-table}
}
\description{
Extract all existing dimension names
}
