% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_helpers.R
\name{get_column_name_all_languages}
\alias{get_column_name_all_languages}
\title{get the names of the fuzzy column in all existing languages}
\usage{
get_column_name_all_languages(df_px, column_name)
}
\arguments{
\item{df_px}{list with the information}

\item{column_name}{column name}
}
\description{
get the names of the fuzzy column in all existing languages
}
