% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dimensions.R
\name{get_dimensions}
\alias{get_dimensions}
\title{Function to import specific dimensions}
\usage{
get_dimensions(unique_dimension_names)
}
\arguments{
\item{unique_dimension_names}{dimension names to import (in english)}
}
\description{
Function to import specific dimensions
}
