% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dimensions.R
\name{get_existing_dimensions}
\alias{get_existing_dimensions}
\title{Function to list all existing dimensions}
\usage{
get_existing_dimensions()
}
\description{
Function to list all existing dimensions
}
