% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_to_dict.R
\name{merge_to_dict}
\alias{merge_to_dict}
\title{Function to merge the values to a dictionary}
\usage{
merge_to_dict(df, var_name, dict)
}
\arguments{
\item{df}{a df where the values are included}

\item{var_name}{the name of the variable to be replaced}

\item{dict}{a df with the value_ids to merge}
}
\description{
This function corrects the values of a column with the supposed data values
in a dictionary.
}
