% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_helpers.R
\name{prep_needed_values}
\alias{prep_needed_values}
\title{get needed values}
\usage{
prep_needed_values(input_list, ignore_dim)
}
\arguments{
\item{input_list}{px file list}

\item{ignore_dim}{dimenstions to be ignored}
}
\description{
get needed values
}
