% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_helpers.R
\name{split_vairable_dimension}
\alias{split_vairable_dimension}
\title{create new dimenstions based on the fuzzy column}
\usage{
split_vairable_dimension(needed_values, fuzzy_column_pattern)
}
\arguments{
\item{needed_values}{list with the needed values}

\item{fuzzy_column_pattern}{pattern containing the fuzzy column name in all existing lanugages}
}
\description{
create new dimenstions based on the fuzzy column
}
