% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_03_03_CH.R
\name{statbot_src_03_03_03_CH}
\alias{statbot_src_03_03_03_CH}
\title{Function do download the dataset with the id: 03_03_03_CH}
\usage{
statbot_src_03_03_03_CH(flag_force_update = FALSE)
}
\arguments{
\item{flag_force_update}{To force an update of the Dataset. Default: FALSE}
}
\description{
Created by: Matthias Widmer
Created at: 2022-02-22
}
