% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_01_001_BS.R
\name{statbot_src_1_01_001_BS}
\alias{statbot_src_1_01_001_BS}
\title{Function do download the dataset with the id: 1_01_001_BS}
\usage{
statbot_src_1_01_001_BS(flag_force_update = FALSE)
}
\arguments{
\item{flag_force_update}{To force an update of the Dataset. Default: FALSE}
}
\description{
Created by: Hester Pieters
Created at: 2022-02-25
}
