% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_01_001_CH.R
\name{statbot_src_1_01_001_CH}
\alias{statbot_src_1_01_001_CH}
\title{Function do download the dataset with the id: 1_01_001_CH}
\usage{
statbot_src_1_01_001_CH(flag_force_update = FALSE)
}
\arguments{
\item{flag_force_update}{To force an update of the Dataset. Default: FALSE}
}
\description{
Created by: Jorin Steiger
Created at: 2022-02-23
Version 3.2.0 - 2022-02-23
}
\details{
History
}
