% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_01_008_CH.R
\name{statbot_src_1_01_008_CH}
\alias{statbot_src_1_01_008_CH}
\title{Function do download the dataset with the id: 1_01_008_CH}
\usage{
statbot_src_1_01_008_CH(flag_force_update = FALSE)
}
\arguments{
\item{flag_force_update}{To force an update of the Dataset. Default: FALSE}
}
\description{
Created by: Christian Ruiz
Created at: 2022-01-06
Version 3.2.0 - 18.02.2022
}
\details{
History

Version 3.2.0 - 18.02.2022 - new dimensions extraction
Version 3.1.2 - 06.01.2022 - First Version
}
