% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_01_011.R
\name{statbot_src_1_01_011_CH}
\alias{statbot_src_1_01_011_CH}
\title{Function do download the dataset with the id: 1_01_011_CH}
\usage{
statbot_src_1_01_011_CH(flag_force_update = FALSE)
}
\arguments{
\item{flag_force_update}{To force an update of the Dataset. Default: FALSE}
}
\description{
Created by: Christian Ruiz
Created at: 2022-03-09
Version 3.2.0 - 09.03.2022
}
\details{
History

Version 3.2.0 - 09.03.2022 - First version copied from 1_01_011
}
