% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_01_027.R
\name{statbot_src_1_01_027}
\alias{statbot_src_1_01_027}
\title{Function do download the dataset with the id: 1_01_027}
\usage{
statbot_src_1_01_027(flag_force_update = FALSE)
}
\arguments{
\item{flag_force_update}{To force an update of the Dataset. Default: FALSE}
}
\description{
Created by: J. Steiger
Created at: 2022-03-14
}
