% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{statbot_src_test}
\alias{statbot_src_test}
\title{Function do download the dataset with the id: test}
\usage{
statbot_src_test(flag_force_update = FALSE)
}
\arguments{
\item{flag_force_update}{To force an update of the Dataset. Default: FALSE}
}
\description{
Created by: Thomas Knecht
Created at: 2021-12-16
}
