% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_to_spatial_unit_name.R
\name{translate_to_spatial_unit_name}
\alias{translate_to_spatial_unit_name}
\title{Function to translate to names of spatial units}
\usage{
translate_to_spatial_unit_name(df, language)
}
\arguments{
\item{df}{a data.frame containing on the one hand an ontology vector called spatialunit_ontology,
and on the other hand a vector called spatialunit_hist_id. Please mind, that it is not spatialunit_current_id.}

\item{language}{a character vector specifiying the language (de,fr,it,en)}
}
\description{
This function translates hist-ids with ontologies to name of spatial units
}
