% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_last_updated.R
\name{update_last_updated}
\alias{update_last_updated}
\title{Function to update the time-stamp in classes function}
\usage{
update_last_updated(id)
}
\arguments{
\item{id}{to be updated}
}
\description{
This function updates the time-stamp in the classes function
}
