% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_meta_and_generate_dimensions.R
\name{extract_meta_and_generate_dimensions}
\alias{extract_meta_and_generate_dimensions}
\title{Function to extract metadata and generate dimensions}
\usage{
extract_meta_and_generate_dimensions(input_df, label, dim_id, ...)
}
\arguments{
\item{input_df}{a px-object before conversion into a proper df}

\item{label}{German label for the variable with first letter capital e.g. "Wirtschaftssektor"}

\item{dim_id}{the unique id in the dimensions table}

\item{...}{dim_name_xx or value_name_xx with corresponding languages. Careful when giving in Labels
they have to be as they would appear with appropriate capital letters e.g. "Economic Sector". The English
one is particular and will be then be automatically converted to e.g. "economic_sector"}
}
\description{
This function extracts metadata from a px-file and populates the dimensions-file
}
